==> 10_scripts/45_eda.py <==
import pandas as pd
import os
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_parquet(
    os.path.join("./20_intermediate_files/W1_countries/ALLCOUNTRIES.parquet")
)
# -*- coding: utf-8 -*-
"""Untitled2.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1UjruGrphIwV1E-CgicSPjq1etPosYBZE
"""

sns.set_style("white")
fig = sns.catplot(
    x="qoi006",
    order=["Very Bad", "Bad", "So so", "Good", "Very Good"],
    hue="country",
    kind="count",
    data=df,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("rate the overall economic condition of our country today?")
fig.axes[0, 0].set_ylabel("Frequency")
# fig.add_legend(loc = "lower center")

# sns.set(rc={"figure.figsize":(5, 10)})
sns.set_style("white")
fig = sns.catplot(
    x="qoi011",
    order=[
        "Much Worse",
        "A little Worse",
        "About the same",
        "A little Better",
        "Much Better",
    ],
    hue="country",
    kind="count",
    data=df,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("Economic situation of your family five years from now?")
fig.axes[0, 0].set_ylabel("Frequency")

sns.set_style("white")
fig = sns.catplot(
    x="qoi050",
    order=[
        "Strongly disagree",
        "Somewhat disagree",
        "Somewhat agree",
        "Strongly agree",
    ],
    hue="country",
    kind="count",
    data=df,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("The military should come in to govern the country")
fig.axes[0, 0].set_ylabel("Frequency")

dfwave2 = pd.read_parquet(
    os.path.join("./20_intermediate_files/W2_countries/ALLCOUNTRIES.parquet")
)

sns.set_style("white")
fig = sns.catplot(
    x="qoi006",
    order=["Very Bad", "Bad", "So so", "Good", "Very Good"],
    hue="country",
    kind="count",
    data=dfwave2,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("rate the overall economic condition of our country today?")
fig.axes[0, 0].set_ylabel("Frequency")

sns.set_style("white")
fig = sns.catplot(
    x="qoi011",
    order=[
        "Much Worse",
        "A little Worse",
        "About the same",
        "A little Better",
        "Much Better",
    ],
    hue="country",
    kind="count",
    data=dfwave2,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("Economic situation of your family five years from now?")
fig.axes[0, 0].set_ylabel("Frequency")

sns.set_style("white")
fig = sns.catplot(
    x="qoi050",
    order=[
        "Strongly disagree",
        "Somewhat disagree",
        "Somewhat agree",
        "Strongly agree",
    ],
    hue="country",
    kind="count",
    data=dfwave2,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("The military should come in to govern the country")
fig.axes[0, 0].set_ylabel("Frequency")


df = pd.read_parquet(
    os.path.join("./20_intermediate_files/W1_countries/Mainland China.parquet")
)

sns.set_style("white")
fig = sns.catplot(
    x="qoi018",
    hue="socialstatus",
    order=[
        "None at all",
        "Not very much trust",
        "A great deal of trust",
        "Quite a lot of trust",
    ],
    data=df,
    kind="count",
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("How much trust do you have in local government?")
fig.axes[0, 0].set_ylabel("Frequency")

sns.set_style("white")
fig = sns.catplot(
    y="qoi041",
    x="agegroup",
    col="married",
    data=df,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel(
    "To what extent would you want our country to be democratic now?"
)
fig.axes[0, 0].set_ylabel("Scale")

dfwave2 = pd.read_parquet(
    os.path.join("./20_intermediate_files/W2_countries/Mainland China.parquet")
)


sns.set_style("white")
fig = sns.catplot(
    x="qoi018",
    hue="socialstatus",
    order=[
        "None at all",
        "Not very much trust",
        "A great deal of trust",
        "Quite a lot of trust",
    ],
    data=df2wave2,
    kind="count",
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel("How much trust do you have in local government?")
fig.axes[0, 0].set_ylabel("Frequency")

sns.set_style("white")
fig = sns.catplot(
    y="qoi041",
    x="agegroup",
    col="married",
    data=df2wave2,
    palette="rocket",
    height=5,
    aspect=1.5,
)
fig.axes[0, 0].set_xlabel(
    "To what extent would you want our country to be democratic now?"
)
fig.axes[0, 0].set_ylabel("Scale")
